refactor(storages): rename storage files to match class names
Storages uses ES6 exports and files are renamed to match the class names.

DEPRECATIONS:
- Old filenames are still supported.
- Module.exports are still supported in OLD FILES.
