+++ src/storages/none.js
@@ -1,35 +1,9 @@
-import Bluebird from 'bluebird';
+import Storage from './Storage';
 
-/**
- * @class NoneStorage
- */
-module.exports = class Storage {
-  /**
-   * Does nothing.
-   *
-   * @param {String} migrationName - Name of migration to be logged.
-   * @returns {Promise}
-   */
-  logMigration(migrationName) {
-    return Bluebird.resolve();
-  }
+module.exports = Storage;
 
-  /**
-   * Does nothing.
-   *
-   * @param {String} migrationName - Name of migration to unlog.
-   * @returns {Promise}
-   */
-  unlogMigration(migrationName) {
-    return Bluebird.resolve();
-  }
-
-  /**
-   * Does nothing.
-   *
-   * @returns {Promise.<String[]>}
-   */
-  executed() {
-    return Bluebird.resolve([]);
-  }
-}
+console.warn(
+  'Deprecated: Storage\'s (former none storage) filename has changed!',
+  'Use \'umzug/lib/storages/Storage\' instead of \'umzug/lib/storages/none\'',
+  'For more information: https://github.com/sequelize/umzug/pull/138',
+);
